pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.13.0",
    "jqdatasdk>=1.8.0",
    "lightgbm>=3.2.0",
    "numpy>=1.19.0",
    "pandas>=1.3.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=5.4.0",
    "scikit-learn>=0.24.0",
]
